+++ meilisearch-http/src/extractors/mod.rs
@@ -1,1 +1,2 @@
 pub mod payload;
+pub mod authentication;
